refactor: 세션, 로컬 저장소 설정 분리
