+++ src/state/localPersist.js
@@ -1,0 +1,6 @@
+import { recoilPersist } from 'recoil-persist';
+
+export const { persistAtom } = recoilPersist({
+  key: 'recoil-persist',
+  storage: localStorage,
+});
