+++ src/state/loginState.js
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import { persistAtom } from './persist';
+import { persistAtom } from './localPersist';
 
 export const loginState = atom({
   key: 'loginState',
